NJ: swap event scrapers out, keep older
